fix: extract server-compatible functions to prevent cross-package type errors
Moved getOrganizationStep and areAllStepsComplete to step-completion-checks.ts
to avoid pulling React components when importing from @propsto/auth/post-auth-check.

This fixes the type-check failure in CI where the app package was trying to
type-check auth's React components due to transitive imports.
